Fix match result bug and allow for user input
I figured out that the bug was in the playRound function. I ditched
using booleans for the playerWin variable with a simple yes/no/tie
string system. That way the computer wouldn't win by default upon a
tied match. It also adds a bit of clarity as I can output this data to
the console for a definite answer to who won. Once I did that, I
adjusted the game function so that the scores would add up properly.

I have a lot of console.log statements to show results. They're a bit
redundant, but I will leave them in for now. They affirm each step I
wrote was successful, like a pat on the back.

I also wrote one line of code that allows for case-insensitive user
input. The user can type rock, PAPER, or sCISsorS and the computer will
convert it to lower caps. There's also an else statement that calls out
the user for using a different word, symbol or number. As a bonus,
it won't increment either the player or computer score. It concludes
with a tie if the player refuses to play properly.

I also cleaned up the syntax a bit for readability. Took away, renamed
 and rearranged everything I could.

 I guess the next step is either sharing right away or making the game
 prettier to look at with UI. But that's a little ways off. I suppose I
 could keep looking for ways to clean up the code and make it even more
 efficient. Also, add more comments (remember write why not how).