Create buttons with event listeners
I removed the game() function from the code to focus on implementing
the events. The game will no longer loop until five rounds are complete.

Now each button, upon clicking, will play a round of
rock paper scissors (the button corresponds to the player input). The
results are then displayed in the browser, not the console.

There is a lot of repeated code. A callback functions should handle
that.