Create Scoreboard
Now when the a button is clicked, text appears on the browser displaying
the result of the match. Player or computer score is incremented
depending on winner.

Right now the logic is so that when player score + comp score is equal
to 3, the game displays the final winner of the game. However the player
is allowed to keep playing. Also the game does not take tied matches
into consideration.

-Create a 'tie' counter to limit the number of matches to 5 per game
-Try to make match results more readable
-Create a reset button after game ends, sets counter back to 0 and
clears results div